refactor: FoodSquareCard・FavoriteStarをコンポーネントとして抽出（#20）
